Move worldinfo endpoints into separate file
